page_object/requirements.txt
selenium
allure-pytest==2.13.5
